pyproject.toml
[project]
name = "hyena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-core>=0.14.13",
    "llama-index-embeddings-openai>=0.5.1",
    "llama-index-readers-smart-pdf-loader>=0.4.1",
    "llama-parse>=0.6.92",
    "llmsherpa>=0.1.4",
    "python-dotenv>=1.2.1",
]
